Major bugfix -- see details
Adjusted the write buffer and miss interaction to fulfill the project requirements correctly.

Specifically, any cache misses need to fetch from L2. As such, they must wait for the write buffer to finish if it is in transfer progress. They now properly take the penalty immediately. Instead of setting l2_used to 0 as I was doing before to address this and it broke everything, it needs to be set to 1. This signals to the write buffer that it must wait until the next cycle to begin writing again because the memory was used this cycle and that (hopefully) correctly resolves the conflict.

Now I have to redo the experiment 1 though.